Internal #861: Refactor WindowAggregateStates
Move the WindowAggregateStates class to separate files.
